refactor(metric): move query logic to repository layer, extract shared types
- Move Prometheus query pipeline (PromQL build + call + parse) from
  UtilizationMetricService to MetricRepository
- Extract MetricValue/ValueType to common/clients/prometheus/types.py
- Move UtilizationMetricType, DIFF/RATE_METRICS to data/metric/types.py
- Move CONTAINER_UTILIZATION_METRIC_NAME to common/metrics/types.py
- Add KernelMetricValuesByKernel to repositories/metric/types.py
- Remove MetricRepositories (MetricRepository created directly in factory)
- Service now delegates to repository, assembles batch result only
- Each file imports directly from the defining module (no re-exports)
