+++ src/ai/backend/common/clients/prometheus/__init__.py
@@ -1,11 +1,12 @@
 from .client import PrometheusClient
 from .preset import LabelMatcher, LabelOperator, MetricPreset
 from .querier import ContainerMetricQuerier, MetricQuerier
-from .types import ValueType
+from .types import MetricValue, ValueType
 
 __all__ = [
     "LabelMatcher",
     "LabelOperator",
+    "MetricValue",
     "PrometheusClient",
     "MetricPreset",
     "MetricQuerier",
