+++ src/ai/backend/common/clients/prometheus/types.py
@@ -1,3 +1,4 @@
+from dataclasses import dataclass
 from enum import StrEnum
 
 
@@ -6,3 +7,13 @@ class ValueType(StrEnum):
 
     CURRENT = "current"
     CAPACITY = "capacity"
+    PCT = "pct"
+
+
+@dataclass(frozen=True)
+class MetricValue:
+    """A single metric sample with its name, value type, and raw value."""
+
+    metric_name: str
+    value_type: ValueType
+    value: str
