+++ src/ai/backend/manager/services/metric/root_service.py
@@ -1,52 +1,24 @@
-import asyncio
 import logging
-import re
-from collections.abc import Sequence
-from typing import Final
 
 from ai.backend.common.clients.prometheus.client import PrometheusClient
-from ai.backend.common.clients.prometheus.preset import LabelMatcher, MetricPreset
 from ai.backend.common.exception import (
     FailedToGetMetric,
     PrometheusConnectionError,
-    UnreachableError,
 )
-from ai.backend.common.metrics.types import UTILIZATION_METRIC_INTERVAL
-from ai.backend.common.types import KernelId
 from ai.backend.logging import BraceStyleAdapter
+from ai.backend.manager.data.metric.types import KernelLiveStatBatchResult
 from ai.backend.manager.repositories.metric.repository import MetricRepository
 
 from .actions.live_stat import QueryKernelLiveStatAction, QueryKernelLiveStatActionResult
 from .container_metric import (
-    CONTAINER_UTILIZATION_METRIC_NAME,
     ContainerUtilizationMetricService,
-)
-from .types import (
-    DIFF_METRICS,
-    RATE_METRICS,
-    KernelLiveStatBatchResult,
-    KernelMetricValuesByKernel,
-    UtilizationMetricType,
-    ValueType,
 )
 
 log = BraceStyleAdapter(logging.getLogger(__spec__.name))
 
-_LIVE_STAT_GROUP_BY: Final[frozenset[str]] = frozenset({
-    "kernel_id",
-    "container_metric_name",
-    "value_type",
-})
-
-
-def _regex_union(values: Sequence[str]) -> str:
-    return "|".join(re.escape(value) for value in values)
-
 
 class UtilizationMetricService:
     container: ContainerUtilizationMetricService
-    _prometheus_client: PrometheusClient
-    _timewindow: str
     _metric_repository: MetricRepository
 
     def __init__(
@@ -55,8 +27,6 @@ class UtilizationMetricService:
         timewindow: str,
         metric_repository: MetricRepository,
     ) -> None:
-        self._prometheus_client = prometheus_client
-        self._timewindow = timewindow
         self.container = ContainerUtilizationMetricService(prometheus_client, timewindow=timewindow)
         self._metric_repository = metric_repository
 
@@ -69,23 +39,8 @@ class UtilizationMetricService:
                 stats=KernelLiveStatBatchResult.empty(action.kernel_ids)
             )
         try:
-            gauge, diff, rate = await asyncio.gather(
-                self._query_kernel_live_stat(
-                    action.kernel_ids,
-                    metric_type=UtilizationMetricType.GAUGE,
-                ),
-                self._query_kernel_live_stat(
-                    action.kernel_ids,
-                    metric_type=UtilizationMetricType.DIFF,
-                    metric_name_filter=DIFF_METRICS,
-                    value_type_filter=ValueType.CURRENT,
-                ),
-                self._query_kernel_live_stat(
-                    action.kernel_ids,
-                    metric_type=UtilizationMetricType.RATE,
-                    metric_name_filter=RATE_METRICS,
-                    value_type_filter=ValueType.CURRENT,
-                ),
+            values_by_kernel = await self._metric_repository.query_kernel_live_stats(
+                action.kernel_ids,
             )
         except (PrometheusConnectionError, FailedToGetMetric):
             log.warning("Failed to query Prometheus for kernel live stats, returning empty results")
@@ -93,68 +48,9 @@ class UtilizationMetricService:
                 stats=KernelLiveStatBatchResult.empty(action.kernel_ids)
             )
 
-        merged = gauge.merged_with(diff).merged_with(rate)
         return QueryKernelLiveStatActionResult(
             stats=KernelLiveStatBatchResult.from_metric_values(
                 action.kernel_ids,
-                merged.values_by_kernel,
+                values_by_kernel,
             )
-        )
-
-    async def _query_kernel_live_stat(
-        self,
-        kernel_ids: Sequence[KernelId],
-        *,
-        metric_type: UtilizationMetricType,
-        metric_name_filter: frozenset[str] | None = None,
-        value_type_filter: ValueType | None = None,
-    ) -> KernelMetricValuesByKernel:
-        preset = self._build_live_stat_preset(
-            kernel_ids,
-            metric_type=metric_type,
-            metric_name_filter=metric_name_filter,
-            value_type_filter=value_type_filter,
-        )
-        response = await self._prometheus_client.query_instant(preset)
-        return KernelMetricValuesByKernel.from_prometheus_response(response)
-
-    def _build_live_stat_preset(
-        self,
-        kernel_ids: Sequence[KernelId],
-        *,
-        metric_type: UtilizationMetricType,
-        metric_name_filter: frozenset[str] | None = None,
-        value_type_filter: ValueType | None = None,
-    ) -> MetricPreset:
-        # TODO: Metrics repository should be used here to dynamically fetch metric names and label values instead of hardcoding regexes
-        labels: dict[str, LabelMatcher] = {
-            "kernel_id": LabelMatcher.regex(_regex_union([str(kid) for kid in kernel_ids]))
-        }
-        if metric_name_filter is not None:
-            labels["container_metric_name"] = LabelMatcher.regex(
-                _regex_union(sorted(metric_name_filter))
-            )
-        if value_type_filter is not None:
-            labels["value_type"] = LabelMatcher.exact(value_type_filter.value)
-
-        match metric_type:
-            case UtilizationMetricType.GAUGE:
-                template = (
-                    f"sum by ({{group_by}})({CONTAINER_UTILIZATION_METRIC_NAME}{{{{{{labels}}}}}})"
-                )
-            case UtilizationMetricType.RATE:
-                template = (
-                    "sum by ({group_by})(rate("
-                    f"{CONTAINER_UTILIZATION_METRIC_NAME}{{{{{{labels}}}}}}[{{window}}]))"
-                    f" / {UTILIZATION_METRIC_INTERVAL}"
-                )
-            case UtilizationMetricType.DIFF:
-                template = f"sum by ({{group_by}})(rate({CONTAINER_UTILIZATION_METRIC_NAME}{{{{{{labels}}}}}}[{{window}}]))"
-            case _:
-                raise UnreachableError(f"Unsupported metric type: {metric_type}")
-        return MetricPreset(
-            template=template,
-            labels=labels,
-            group_by=_LIVE_STAT_GROUP_BY,
-            window=self._timewindow,
         )
