+++ src/ai/backend/manager/services/metric/types.py
@@ -1,30 +1,16 @@
 from __future__ import annotations
 
-import enum
 from dataclasses import dataclass
-from typing import Final, Self, cast
+from typing import Self
 from uuid import UUID
 
 from pydantic import BaseModel
 
+from ai.backend.common.clients.prometheus.types import ValueType
 from ai.backend.common.dto.clients.prometheus.response import (
     MetricResponseInfo,
-    PrometheusResponse,
 )
 from ai.backend.common.exception import InvalidAPIParameters
-from ai.backend.common.types import KernelId
-from ai.backend.manager.data.metric.types import (
-    KernelLiveStatBatchResult as KernelLiveStatBatchResult,
-)
-from ai.backend.manager.data.metric.types import (
-    KernelLiveStatEntry as KernelLiveStatEntry,
-)
-from ai.backend.manager.data.metric.types import (
-    KernelMetricValue,
-)
-from ai.backend.manager.data.metric.types import (
-    ValueType as ValueType,
-)
 
 
 class MetricQueryParameter(BaseModel):
@@ -87,72 +73,3 @@ class ContainerMetricOptionalLabel:
 class ContainerMetricResult:
     metric: ContainerMetricResponseInfo
     values: list[MetricResultValue]
-
-
-@dataclass(frozen=True)
-class KernelMetricValuesByKernel:
-    values_by_kernel: dict[KernelId, list[KernelMetricValue]]
-
-    @classmethod
-    def from_prometheus_response(cls, response: PrometheusResponse) -> Self:
-        grouped: dict[KernelId, list[KernelMetricValue]] = {}
-        for metric in response.data.result:
-            info = metric.metric
-            if not info.has_container_metric_labels or not metric.values:
-                continue
-            # Non-None guaranteed by has_container_metric_labels above;
-            # cast needed because property checks don't narrow types.
-            kernel_id_str = cast(str, info.kernel_id)
-            container_metric_name = cast(str, info.container_metric_name)
-            value_type_str = cast(str, info.value_type)
-            try:
-                value_type = ValueType(value_type_str)
-                kernel_id = KernelId(UUID(kernel_id_str))
-            except ValueError:
-                continue
-            # Instant queries are normalized into a one-element list, and range
-            # queries are ordered by time, so the last sample is the newest one.
-            _, raw_value = metric.values[-1]
-            grouped.setdefault(kernel_id, []).append(
-                KernelMetricValue(
-                    metric_name=container_metric_name,
-                    value_type=value_type,
-                    value=raw_value,
-                )
-            )
-        return cls(values_by_kernel=grouped)
-
-    def merged_with(self, other: Self) -> Self:
-        merged = {kernel_id: list(values) for kernel_id, values in self.values_by_kernel.items()}
-        for kernel_id, values in other.values_by_kernel.items():
-            merged.setdefault(kernel_id, []).extend(values)
-        return type(self)(values_by_kernel=merged)
-
-
-class UtilizationMetricType(enum.Enum):
-    """
-    Specifies the type of a metric value.
-    """
-
-    GAUGE = enum.auto()
-    """
-    Represents an instantly measured occupancy value.
-    (e.g., used space as bytes, occupied amount as the number of items or a bandwidth)
-    """
-    RATE = enum.auto()
-    """
-    Represents a rate of changes calculated from underlying gauge/accumulation values
-    (e.g., I/O bps calculated from RX/TX accum.bytes)
-    """
-    DIFF = enum.auto()
-    """
-    Represents a difference of changes calculated from underlying gauge/accumulation values
-    (e.g., Utilization msec from CPU usage)
-    """
-
-
-# Metric-name -> UtilizationMetricType classification rules.
-# TODO: Refactor to query metric metadata from the repository layer once
-#       the metadata persistence is available.
-DIFF_METRICS: Final[frozenset[str]] = frozenset({"cpu_util"})
-RATE_METRICS: Final[frozenset[str]] = frozenset({"net_rx", "net_tx"})
